requirements.txt
requests
tweepy
schedule
huggingface_hub
transformers
torch
accelerate
atproto
python-dotenv